BE/requirements.txt
Django==5.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
django-environ==0.12.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
django-q2==1.7.3
google-generativeai==0.8.4
pytest==8.3.5
pytest-django==4.10.0
factory-boy==3.3.3
